+++ CHANGELOG.md
@@ -20,6 +20,9 @@ All notable changes to this project will be documented in this file.
 
 ## [0.7.3] - To be released
 
+### Added
+- PasswordField.
+
 ## [0.7.2] - 14.08.2022
 
 ### Added