+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.7.3] - To be released
+## [0.7.3] - 31.08.2022
 
 ### Added
 - Spectator join feature in bgw-net.